Moved generic extensions to more appropriate location
