+++ PubNubSwiftConsole/Classes/Extensions/ViewExtensions.swift
@@ -1,0 +1,22 @@
+//
+//  ViewExtensions.swift
+//  Pods
+//
+//  Created by Jordan Zucker on 8/17/16.
+//
+//
+
+import UIKit
+
+extension UIView {
+    
+    var hasConstraints: Bool {
+        let hasHorizontalConstraints = !self.constraintsAffectingLayout(for: .horizontal).isEmpty
+        let hasVerticalConstraints = !self.constraintsAffectingLayout(for: .vertical).isEmpty
+        return hasHorizontalConstraints || hasVerticalConstraints
+    }
+    
+    func forceAutoLayout() {
+        self.translatesAutoresizingMaskIntoConstraints = false
+    }
+}
