+++ PubNubSwiftConsole/Classes/UI/CollectionView/Cells/ResultCollectionViewCell.swift
@@ -44,17 +44,6 @@ class Result: ResultItem {
     }
 }
 
-extension UIView {
-    var hasConstraints: Bool {
-        let hasHorizontalConstraints = !self.constraintsAffectingLayout(for: .horizontal).isEmpty
-        let hasVerticalConstraints = !self.constraintsAffectingLayout(for: .vertical).isEmpty
-        return hasHorizontalConstraints || hasVerticalConstraints
-    }
-    func forceAutoLayout() {
-        self.translatesAutoresizingMaskIntoConstraints = false
-    }
-}
-
 class ResultCollectionViewCell: CollectionViewCell {
     let operationLabel: UILabel
     let creationDateLabel: UILabel
